Fix marking TODOs as done
For some reason getting TODOs with a group ID now produces a CloudFlare
access check. In another case it produces an internal server error. In
addition, the API call to get the group ID is really slow.

To work around this, we now just match the start of the project
`path_with_namespace` field.